[refactor] Extract ``dummy_highlighter`` to ``_highlight.py``
